// apps/desktop/src/generated/query/TimelineGroupDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TimelineItemDto } from "./TimelineItemDto";
import type { TimelineOriginDto } from "./TimelineOriginDto";

export type TimelineGroupDto = { group_id: string, origin: TimelineOriginDto, occurred_at: string, items: Array<TimelineItemDto>, };